+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 #### Fixed
 
 - Xargo now respects the build.rustflags value set in .cargo/config.
+- A bug where the hash/date file didn't get properly truncated before updating it leading to Xargo
+to *always* trigger a sysroot rebuild.
 
 ### v0.1.2 - 2016-04-24